requirements.txt
pyyaml
requests
pyaudio
webrtcvad
simpleavs
